+++ ChangeLog.md
@@ -8,6 +8,7 @@ Copyright (C) 2012, 2019, 2023 David M. Syzdek <david@syzdek.net>
 ---
   - project: updating to use POSIX C source 200112L (syzdek)
   - project: updating to use C dialect c11 (syzdek)
+  - ldapppolicy: adding utility (syzdek)
 
 0.6
 ---